Add files via upload
MSLattice is a software that can be used to generate the geometries of various types of lattices that are known as periodic cellular materials or architected materials. In particular, lattices based on triply periodic minimal surfaces. The user can generate the STL files needed for fabricating the lattices using additive manufacturing (commonly known as 3D printing) and can generate the files needed for conducting finite element simulations or computational fluid dynamics simulation in order to estimate the performance of lattices under different loading conditions.